(MODULES-1977) Convert Linux provider to POSIX
