+++ acceptance/tests/posix/reboot_immediate.rb
@@ -1,4 +1,4 @@
-test_name "Reboot Module - Linux Provider - Reboot Immediately Explicit"
+test_name "Reboot Module - POSIX Provider - Reboot Immediately"
 extend Puppet::Acceptance::Reboot
 
 reboot_manifest = <<-MANIFEST
@@ -6,14 +6,13 @@ notify { 'step_1':
 }
 ~>
 reboot { 'now':
-  apply => immediately
 }
 MANIFEST
 
 confine :except, :platform => 'windows'
 
-linux_agents.each do |agent|
-  step "Reboot Immediately (Explicit) with Refresh"
+posix_agents.each do |agent|
+  step "Reboot Immediately with Refresh"
 
   #Apply the manifest.
   on agent, puppet('apply', '--debug'), :stdin => reboot_manifest
