+++ acceptance/tests/posix/reboot_message.rb
@@ -1,4 +1,4 @@
-test_name "Reboot Module - Linux Provider - Custom Message"
+test_name "Reboot Module - POSIX Provider - Custom Message"
 extend Puppet::Acceptance::Reboot
 
 reboot_manifest = <<-MANIFEST
@@ -13,12 +13,18 @@ MANIFEST
 
 confine :except, :platform => 'windows'
 
-linux_agents.each do |agent|
+posix_agents.each do |agent|
   step "Reboot Immediately with a Custom Message"
 
   #Apply the manifest.
   on agent, puppet('apply', '--debug'), :stdin => reboot_manifest do |result|
-    assert_match /shutdown -r \+1 \"A different message\"/,
+    case fact('kernel')
+    when 'SunOS'
+      expected_command = /shutdown -y -i 6 -g 60 \"A different message\"/
+    else
+      expected_command = /shutdown -r \+1 \"A different message\"/
+    end
+    assert_match expected_command,
       result.stdout, 'Expected reboot message is incorrect'
   end
