+++ acceptance/tests/posix/reboot_timeout.rb
@@ -1,0 +1,31 @@
+test_name "Reboot Module - POSIX Provider - Custom Timeout"
+extend Puppet::Acceptance::Reboot
+
+reboot_manifest = <<-MANIFEST
+notify { 'step_1':
+}
+~>
+reboot { 'now':
+  when => refreshed,
+  timeout => 120,
+}
+MANIFEST
+
+#Solaris specifies reboot timeout in seconds
+confine :except, :platform => ['windows', 'solaris']
+
+posix_agents.each do |agent|
+  step "Reboot Immediately with a Custom Timeout"
+
+  #Apply the manifest.
+  on agent, puppet('apply', '--debug'), :stdin => reboot_manifest do |result|
+    assert_match /shutdown -r \+2/,
+      result.stdout, 'Expected reboot timeout is incorrect'
+  end
+
+  #Waiting 61 seconds guarantees that the default timeout is different.
+  sleep 61
+
+  #Verify that a shutdown has been initiated and clear the pending shutdown.
+  retry_shutdown_abort(agent)
+end
